sdm660-common: manifest: Register android.hardware.health@2.1 HAL
* Get rid of the follwoing logspam
10-08 01:50:41.552 W/HidlServiceManagement(752): Waited one second for
android.hardware.health@2.1::IHealth/default
10-08 01:50:41.555 I/hwservicemanager(483): Since
android.hardware.health@2.1::IHealth/default is not registered, trying
to start it as a lazy HAL.
10-08 01:50:40.552 I/hwservicemanager(483): Tried to start
android.hardwar getService: Trying again for
android.hardware.health@2.1::IHealth/default...
10-08 01:50:40.552 W/libc    (483): Unable to set property
"ctl.interface_start" to "android.hardware.health@2.1::IHealth/default":
error code: 0x20e.health@2.1::IHealth/default as a lazy service, but was
unable to. Usually this happens when a service is not installed, but if
the service is intended to be used as a lazy service, then it may be
configured incorrectly.